Split some files to fit the MOBILEAPP case better
