+++ android/lib/src/main/cpp/CMakeLists.txt.in
@@ -10,6 +10,7 @@ add_library(androidapp SHARED
             ../../../../../common/CommandControl.cpp
             ../../../../../common/Crypto-stub.cpp
             ../../../../../common/FileUtil.cpp
+            ../../../../../common/FileUtil-unix.cpp
             ../../../../../common/Log.cpp
             ../../../../../common/TraceEvent.cpp
             ../../../../../common/Protocol.cpp
@@ -17,11 +18,12 @@ add_library(androidapp SHARED
             ../../../../../common/Simd.cpp
             ../../../../../common/StringVector.cpp
             ../../../../../common/Session.cpp
-            ../../../../../common/SigUtil.cpp
+            ../../../../../common/SigUtil-mobile.cpp
             ../../../../../common/SpookyV2.cpp
             ../../../../../common/Unit.cpp
             ../../../../../common/Util.cpp
             ../../../../../common/Util-mobile.cpp
+            ../../../../../common/Util-unix.cpp
             ../../../../../common/Uri.cpp
             ../../../../../kit/ChildSession.cpp
             ../../../../../kit/DeltaSimd.c
