+++ common/FileUtil.cpp
@@ -18,22 +18,12 @@
 #include <common/Unit.hpp>
 #include <common/Util.hpp>
 
-#include <dirent.h>
 #include <exception>
-#include <ftw.h>
 #include <grp.h>
 #include <pwd.h>
 #include <stdexcept>
 #include <sys/time.h>
 #include <unistd.h>
-#ifdef __linux__
-#include <sys/vfs.h>
-#elif defined IOS
-#import <Foundation/Foundation.h>
-#elif defined __FreeBSD__
-#include <sys/param.h>
-#include <sys/mount.h>
-#endif
 
 #include <fcntl.h>
 #include <chrono>
@@ -54,7 +44,7 @@ namespace FileUtil
     std::string createRandomDir(const std::string& path)
     {
         std::string name = Util::rng::getFilename(64);
-        std::filesystem::create_directory(path + '/' + name);
+        createDirectory(path + '/' + name);
         return name;
     }
 
@@ -81,7 +71,7 @@ namespace FileUtil
         int from = -1, to = -1;
         try
         {
-            from = open(fromPath.c_str(), O_RDONLY);
+            from = openFileAsFD(fromPath, O_RDONLY);
             if (from < 0)
                 throw std::runtime_error("Failed to open src " + anonymizeUrl(fromPath));
 
@@ -89,7 +79,7 @@ namespace FileUtil
             if (fstat(from, &st) != 0)
                 throw std::runtime_error("Failed to fstat src " + anonymizeUrl(fromPath));
 
-            to = open(toPath.c_str(), O_CREAT | O_TRUNC | O_WRONLY, st.st_mode);
+            to = openFileAsFD(toPath, O_CREAT | O_TRUNC | O_WRONLY, st.st_mode);
             if (to < 0)
                 throw std::runtime_error("Failed to open dest " + anonymizeUrl(toPath));
 
@@ -128,8 +118,8 @@ namespace FileUtil
                 LOG_WRN("Unusual: file " << anonymizeUrl(fromPath) << " changed size "
                         "during copy from " << st.st_size << " to " << bytesIn);
             }
-            close(from);
-            close(to);
+            closeFD(from);
+            closeFD(to);
             return true;
         }
         catch (const std::exception& ex)
@@ -139,8 +129,8 @@ namespace FileUtil
                 << anonymizeUrl(toPath) << ": " << ex.what();
             const std::string err = oss.str();
             LOG_ERR(err);
-            close(from);
-            close(to);
+            closeFD(from);
+            closeFD(to);
             unlink(toPath.c_str());
             if (throw_on_error)
                 throw std::runtime_error(err);
@@ -200,115 +190,6 @@ namespace FileUtil
             return root;
         }
         return newTmp;
-    }
-
-#if 1 // !HAVE_STD_FILESYSTEM
-    static int nftw_cb(const char *fpath, const struct stat*, int type, struct FTW*)
-    {
-        if (type == FTW_DP)
-        {
-            rmdir(fpath);
-        }
-        else if (type == FTW_F || type == FTW_SL)
-        {
-            unlink(fpath);
-        }
-
-        // Always continue even when things go wrong.
-        return 0;
-    }
-#endif
-
-    void removeFile(const std::string& path, const bool recursive)
-    {
-        LOG_DBG("Removing [" << Anonymizer::anonymizeUrl(path) << "] " << (recursive ? "recursively." : "only."));
-
-// Amazingly filesystem::remove_all silently fails to work on some
-// systems. No real need to be using experimental API here either.
-#if 0 // HAVE_STD_FILESYSTEM
-        std::error_code ec;
-        if (recursive)
-            std::filesystem::remove_all(path, ec);
-        else
-            std::filesystem::remove(path, ec);
-
-        // Already removed or we don't care about failures.
-        (void) ec;
-#else
-        try
-        {
-            struct stat sb;
-            errno = 0;
-            if (!recursive || stat(path.c_str(), &sb) == -1 || S_ISREG(sb.st_mode))
-            {
-                // Non-recursive directories and files that exist.
-                if (errno != ENOENT)
-                    Poco::File(path).remove(recursive);
-            }
-            else
-            {
-                // Directories only.
-                nftw(path.c_str(), nftw_cb, 128, FTW_DEPTH | FTW_PHYS);
-            }
-        }
-        catch (const std::exception& e)
-        {
-            // Don't complain if already non-existent.
-            if (FileUtil::Stat(path).exists())
-            {
-                // Error only if it still exists.
-                LOG_ERR("Failed to remove ["
-                        << Anonymizer::anonymizeUrl(path) << "] " << (recursive ? "recursively: " : "only: ") << e.what());
-            }
-        }
-#endif
-    }
-
-    /// Remove directories only, which must be empty for this to work.
-    static int nftw_rmdir_cb(const char* fpath, const struct stat*, int type, struct FTW*)
-    {
-        if (type == FTW_DP)
-        {
-            rmdir(fpath);
-        }
-
-        // Always continue even when things go wrong.
-        return 0;
-    }
-
-    void removeEmptyDirTree(const std::string& path)
-    {
-        LOG_DBG("Removing empty directories at [" << path << "] recursively");
-
-        nftw(path.c_str(), nftw_rmdir_cb, 128, FTW_DEPTH | FTW_PHYS);
-    }
-
-    std::string realpath(const char* path)
-    {
-        char* resolved = ::realpath(path, nullptr);
-        if (resolved)
-        {
-            std::string real = resolved;
-            free(resolved);
-            return real;
-        }
-
-        LOG_SYS("Failed to get the realpath of [" << path << ']');
-        return path;
-    }
-
-    bool isEmptyDirectory(const char* path)
-    {
-        DIR* dir = opendir(path);
-        if (dir == nullptr)
-            return errno != EACCES; // Assume it's not empty when EACCES.
-
-        int count = 0;
-        while (readdir(dir) && ++count < 3)
-            ;
-
-        closedir(dir);
-        return count <= 2; // Discounting . and ..
     }
 
     bool isWritable(const char* path)
@@ -380,11 +261,13 @@ namespace FileUtil
 
     bool compareFileContents(const std::string& rhsPath, const std::string& lhsPath)
     {
-        std::ifstream rhs(rhsPath, std::ifstream::binary | std::ifstream::ate);
+        std::ifstream rhs;
+        openFileToIFStream(rhsPath, rhs, std::ifstream::binary | std::ifstream::ate);
         if (rhs.fail())
             return false;
 
-        std::ifstream lhs(lhsPath, std::ifstream::binary | std::ifstream::ate);
+        std::ifstream lhs;
+        openFileToIFStream(lhsPath, lhs, std::ifstream::binary | std::ifstream::ate);
         if (lhs.fail())
             return false;
 
@@ -580,35 +463,7 @@ namespace FileUtil
         constexpr int64_t ENOUGH_SPACE = gb*1024*1024*1024;
 #endif
 
-#if defined(__linux__) || defined(__FreeBSD__)
-        struct statfs sfs;
-        if (statfs(path.c_str(), &sfs) == -1)
-        {
-            LOG_SYS("Failed to stat filesystem [" << path << ']');
-            return true; // We assume the worst.
-        }
-
-        const int64_t freeBytes = static_cast<int64_t>(sfs.f_bavail) * sfs.f_bsize;
-
-        LOG_INF("Filesystem [" << path << "] has " << (freeBytes / 1024 / 1024) <<
-                " MB free (" << (sfs.f_bavail * 100. / sfs.f_blocks) << "%).");
-
-        if (freeBytes > ENOUGH_SPACE)
-            return true;
-
-        if (static_cast<double>(sfs.f_bavail) / sfs.f_blocks <= 0.05)
-            return false;
-#elif defined IOS
-        NSDictionary *atDict = [[NSFileManager defaultManager] attributesOfFileSystemForPath:@"/" error:NULL];
-        long long freeSpace = [[atDict objectForKey:NSFileSystemFreeSize] longLongValue];
-        long long totalSpace = [[atDict objectForKey:NSFileSystemSize] longLongValue];
-
-        if (freeSpace > ENOUGH_SPACE)
-            return true;
-
-        if (static_cast<double>(freeSpace) / totalSpace <= 0.05)
-            return false;
-#endif
+        return platformDependentCheckDiskSpace(path, ENOUGH_SPACE);
 
         return true;
     }
@@ -628,201 +483,6 @@ namespace FileUtil
         return Util::splitLast(path, '.', true).second;
     }
 
-    void lslr(const std::string& path)
-    {
-        std::cout << path << ":\n";
-
-        DIR* dir = opendir(path.c_str());
-        if (dir == nullptr)
-        {
-            std::cerr << "lslr: fail to open: " << dir << " error: " << std::strerror(errno) << std::endl;
-            return;
-        }
-
-        struct sb
-        {
-            mode_t _mode;
-            nlink_t _nlink;
-            std::string _uid;
-            std::string _gid;
-            off_t _size;
-            time_t _mtime;
-            std::string _name;
-
-            sb(mode_t mode, nlink_t nlink, std::string uid, std::string gid, off_t size, time_t mtime, std::string name)
-                : _mode(mode)
-                , _nlink(nlink)
-                , _uid(std::move(uid))
-                , _gid(std::move(gid))
-                , _size(size)
-                , _mtime(mtime)
-                , _name(std::move(name))
-            {
-            }
-        };
-
-        std::vector<sb> entries;
-        std::vector<std::string> subdirs;
-        size_t nlink_len = 0;
-        size_t size_len = 0;
-        size_t uid_len = 0;
-        size_t gid_len = 0;
-        size_t blocks = 0;
-
-        while (const dirent* f = readdir(dir))
-        {
-            std::string fullpath(path);
-            if (!fullpath.ends_with("/"))
-                fullpath.push_back('/');
-            fullpath.append(f->d_name);
-
-            struct stat statbuf;
-            if (lstat(fullpath.c_str(), &statbuf) != 0)
-            {
-                std::cerr << "lslr: fail to lstat: " << fullpath << " error: " << std::strerror(errno) << std::endl;
-                continue;
-            }
-
-            size_len = std::max(size_len, std::to_string(statbuf.st_size).size());
-            nlink_len = std::max(nlink_len, std::to_string(statbuf.st_nlink).size());
-
-            std::string uid;
-            struct passwd *pwd = getpwuid(statbuf.st_uid);
-            if (pwd && pwd->pw_name)
-                uid = pwd->pw_name;
-            else
-                uid = std::to_string(statbuf.st_uid);
-            uid_len = std::max(uid_len, uid.size());
-
-            std::string gid;
-            struct group *grp = getgrgid(statbuf.st_gid);
-            if (grp && grp->gr_name)
-                gid = grp->gr_name;
-            else
-                gid = std::to_string(statbuf.st_gid);
-
-            entries.emplace_back(statbuf.st_mode, statbuf.st_nlink, uid, gid, statbuf.st_size, statbuf.st_mtime, f->d_name);
-
-            if (strcmp(f->d_name, ".") != 0 && strcmp(f->d_name, "..") != 0 && (statbuf.st_mode & S_IFMT) == S_IFDIR)
-                subdirs.push_back(std::move(fullpath));
-
-            blocks += statbuf.st_blocks;
-        }
-
-        std::sort(entries.begin(), entries.end(), [](const auto& lhs, const auto& rhs)
-                  { return strcasecmp(lhs._name.c_str(), rhs._name.c_str()) < 0; });
-        std::sort(subdirs.begin(), subdirs.end(), [](const auto& lhs, const auto& rhs)
-                  { return strcasecmp(lhs.c_str(), rhs.c_str()) < 0; });
-
-        closedir(dir);
-
-        // turn 512 blocks into ls-alike default 1024 byte blocks
-        std::cout << "total " << (blocks + 1) / 2 << "\n";
-
-        for (const auto& entry : entries)
-        {
-            bool symbolic_link = false;
-
-            switch (entry._mode & S_IFMT)
-            {
-                case S_IFREG:
-                    std::cout << '-';
-                    break;
-                case S_IFBLK:
-                    std::cout << 'b';
-                    break;
-                case S_IFCHR:
-                    std::cout << 'c';
-                    break;
-                case S_IFDIR:
-                    std::cout << 'd';
-                    break;
-                case S_IFLNK:
-                    std::cout << 'l';
-                    symbolic_link = true;
-                    break;
-                case S_IFIFO:
-                    std::cout << 'p';
-                    break;
-                case S_IFSOCK:
-                    std::cout << 's';
-                    break;
-                default:
-                    std::cout << '?';
-                    break;
-                break;
-            }
-
-            std::cout << ((entry._mode & S_IRUSR) ? "r" : "-");
-            std::cout << ((entry._mode & S_IWUSR) ? "w" : "-");
-            std::cout << ((entry._mode & S_IXUSR) ? "x" : "-");
-            std::cout << ((entry._mode & S_IRGRP) ? "r" : "-");
-            std::cout << ((entry._mode & S_IWGRP) ? "w" : "-");
-            std::cout << ((entry._mode & S_IXGRP) ? "x" : "-");
-            std::cout << ((entry._mode & S_IROTH) ? "r" : "-");
-            std::cout << ((entry._mode & S_IWOTH) ? "w" : "-");
-            std::cout << ((entry._mode & S_IXOTH) ? "x" : "-");
-
-            std::cout << " " << std::right << std::setw(nlink_len) << entry._nlink;
-
-            std::cout << " " << std::left << std::setw(uid_len) << entry._uid;
-
-            std::cout << " " << std::left << std::setw(gid_len) << entry._gid;
-
-            std::cout << " " << std::right << std::setw(size_len) << entry._size;
-
-            struct tm tm;
-            std::cout << " " << std::put_time(localtime_r(&entry._mtime, &tm), "%F %R");
-
-            std::cout << " " << entry._name;
-
-            if (symbolic_link)
-            {
-                std::string fullpath(path);
-                fullpath.append("/").append(entry._name);
-
-                const std::size_t size = entry._size;
-                std::vector<char> target(size + 1);
-                char* target_data = target.data();
-                const ssize_t read = readlink(fullpath.c_str(), target_data, size);
-                if (read <= 0 || static_cast<std::size_t>(read) > size)
-                    std::cerr << "lslr: fail to read: " << fullpath << " error: " << std::strerror(errno) << std::endl;
-                else
-                {
-                    target_data[read] = '\0';
-                    std::cout << " -> " << target.data();
-                }
-            }
-
-            std::cout << "\n";
-        }
-
-        for (const auto& subdir : subdirs)
-        {
-            std::cout << "\n";
-            lslr(subdir);
-        }
-    }
-
-    std::vector<std::string> getDirEntries(const std::string dirPath)
-    {
-        std::vector<std::string> names;
-        DIR *dir = opendir(dirPath.c_str());
-        if (!dir)
-        {
-            LOG_DBG("Read from non-existent directory " + dirPath);
-            return names;
-        }
-        struct dirent *i;
-        while ((i = readdir(dir)))
-        {
-            if (i->d_name[0] == '.')
-                continue;
-            names.push_back(i->d_name);
-        }
-        closedir(dir);
-        return names;
-    }
 } // namespace FileUtil
 
 /* vim:set shiftwidth=4 softtabstop=4 expandtab: */
