+++ common/Session.hpp
@@ -252,7 +252,7 @@ public:
 
     const std::string& getLang() const { return _lang; }
 
-    const std::string& getTimezone() const { return _timezone; }
+    const std::string& getTimezone() const { return _timeZone; }
 
     bool getHaveDocPassword() const { return _haveDocPassword; }
 
@@ -379,7 +379,7 @@ private:
     std::string _lang;
 
     /// Timezone of the user.
-    std::string _timezone;
+    std::string _timeZone;
 
     /// The form factor of the device where the client is running: desktop, tablet, mobile.
     std::string _deviceFormFactor;
