+++ net/Socket.cpp
@@ -20,8 +20,9 @@
 #include <common/TraceEvent.hpp>
 #include <common/Unit.hpp>
 #include <common/Util.hpp>
-#include <common/Util.hpp>
+#if !MOBILEAPP
 #include <common/Watchdog.hpp>
+#endif
 #include <net/HttpRequest.hpp>
 #include <net/NetUtil.hpp>
 #include <net/ServerSocket.hpp>
@@ -44,21 +45,28 @@
 #include <sstream>
 #include <cstdio>
 #include <string>
-#include <unistd.h>
-#include <sysexits.h>
+
 #include <sys/stat.h>
 #include <sys/types.h>
+
+#include <sysexits.h>
+#include <unistd.h>
 #include <sys/un.h>
+
 #ifdef __FreeBSD__
 #include <sys/ucred.h>
 #endif
 
 #include <Poco/MemoryStream.h>
+#if !MOBILEAPP
 #include <Poco/Net/HTTPRequest.h>
 #include <Poco/Net/HTTPResponse.h>
 #include <Poco/Net/NetException.h>
 #include <Poco/Net/WebSocket.h> // computeAccept
+#endif
+
 #include <Poco/URI.h>
+
 #if ENABLE_SSL
 #include <Poco/Net/X509Certificate.h>
 #endif
@@ -72,6 +80,8 @@ namespace ThreadChecks
     std::atomic<bool> Inhibit(false);
 }
 
+#if !MOBILEAPP
+
 std::unique_ptr<Watchdog> SocketPoll::PollWatchdog;
 
 std::atomic<size_t> StreamSocket::ExternalConnectionCount = 0;
@@ -80,6 +90,8 @@ net::DefaultValues net::Defaults = { .inactivityTimeout = std::chrono::seconds(3
                                      .maxExtConnections = 200000 /* arbitrary value to be resolved */ };
 
 #define SOCKET_ABSTRACT_UNIX_NAME "0coolwsd-"
+
+#endif
 
 constexpr std::string_view Socket::toString(Type t)
 {
@@ -301,7 +313,9 @@ SocketPoll::SocketPoll(std::string threadName)
     , _pollStartIndex(0)
     , _owner(std::this_thread::get_id())
     , _threadStarted(0)
+#if !MOBILEAPP
     , _watchdogTime(Watchdog::getDisableStamp())
+#endif
     , _ownerThreadId(Util::getThreadId())
     , _stop(false)
     , _threadFinished(false)
@@ -309,9 +323,11 @@ SocketPoll::SocketPoll(std::string threadName)
 {
     ProfileZone profileZone("SocketPoll::SocketPoll");
 
+#if !MOBILEAPP
     static bool watchDogProfile = !!getenv("COOL_WATCHDOG");
     if (watchDogProfile && !PollWatchdog)
         PollWatchdog = std::make_unique<Watchdog>();
+#endif
 
     _wakeup[0] = -1;
     _wakeup[1] = -1;
@@ -320,16 +336,20 @@ SocketPoll::SocketPoll(std::string threadName)
 
     LOG_DBG("New " << logInfo());
 
+#if !MOBILEAPP
     if (PollWatchdog)
         PollWatchdog->addTime(&_watchdogTime, &_ownerThreadId);
+#endif
 }
 
 SocketPoll::~SocketPoll()
 {
     LOG_DBG("~" << logInfo());
 
+#if !MOBILEAPP
     if (PollWatchdog)
         PollWatchdog->removeTime(&_watchdogTime);
+#endif
 
     joinThread();
 
@@ -471,12 +491,16 @@ void SocketPoll::pollingThreadEntry()
 
 void SocketPoll::disableWatchdog()
 {
+#if !MOBILEAPP
     _watchdogTime = Watchdog::getDisableStamp();
+#endif
 }
 
 void SocketPoll::enableWatchdog()
 {
+#if !MOBILEAPP
     _watchdogTime = Watchdog::getTimestamp();
+#endif
 }
 
 int SocketPoll::poll(int64_t timeoutMaxMicroS, bool justPoll)
@@ -781,8 +805,11 @@ void SocketPoll::closeAllSockets()
     for (std::shared_ptr<Socket> &it : _pollSockets)
     {
         // first close the underlying socket
+#if !MOBILEAPP
         it->closeFD(*this);
-
+#else
+        fakeSocketClose(it->getFD());
+#endif
         // avoid the socketHandler' getting an onDisconnect
         auto stream = dynamic_cast<StreamSocket *>(it.get());
         if (stream)
@@ -1830,8 +1857,6 @@ bool StreamSocket::sniffSSL() const
             _inBuffer[5] == 0x01);  // Handshake: CLIENT_HELLO
 }
 
-#endif // !MOBILEAPP
-
 namespace {
     /// To make the protected 'computeAccept' accessible.
     class PublicComputeAccept final : public Poco::Net::WebSocket
@@ -1859,6 +1884,8 @@ std::string WebSocketHandler::generateKey()
 {
     return PublicComputeAccept::generateKey();
 }
+
+#endif // !MOBILEAPP
 
 // Required by Android and iOS apps.
 namespace http
