+++ wsd/COOLWSD.cpp
@@ -107,8 +107,12 @@
 #include <wsd/Process.hpp>
 #include <common/JsonUtil.hpp>
 #include <common/FileUtil.hpp>
+
+#if !MOBILEAPP
 #include <common/JailUtil.hpp>
 #include <common/Watchdog.hpp>
+#endif
+
 #include <common/Log.hpp>
 #include <MobileApp.hpp>
 #include <Protocol.hpp>
@@ -1532,10 +1536,10 @@ void COOLWSD::innerInitialize(Poco::Util::Application& self)
             {
                 fprintf(TraceEventFile, "[\n");
                 // Output a metadata event that tells that this is the WSD process
-                fprintf(TraceEventFile, "{\"name\":\"process_name\",\"ph\":\"M\",\"args\":{\"name\":\"WSD\"},\"pid\":%d,\"tid\":%ld},\n",
-                        getpid(), (long) Util::getThreadId());
-                fprintf(TraceEventFile, "{\"name\":\"thread_name\",\"ph\":\"M\",\"args\":{\"name\":\"Main\"},\"pid\":%d,\"tid\":%ld},\n",
-                        getpid(), (long) Util::getThreadId());
+                fprintf(TraceEventFile, "{\"name\":\"process_name\",\"ph\":\"M\",\"args\":{\"name\":\"WSD\"},\"pid\":%ld,\"tid\":%ld},\n",
+                        Util::getProcessId(), Util::getThreadId());
+                fprintf(TraceEventFile, "{\"name\":\"thread_name\",\"ph\":\"M\",\"args\":{\"name\":\"Main\"},\"pid\":%ld,\"tid\":%ld},\n",
+                        Util::getProcessId(), Util::getThreadId());
             }
         }
     }
@@ -1731,7 +1735,7 @@ void COOLWSD::innerInitialize(Poco::Util::Application& self)
         CleanupChildRoot = ChildRoot;
 
         // Encode the process id into the path for parallel re-use of jails/
-        ChildRoot += std::to_string(getpid()) + '-' + Util::rng::getHexString(8) + '/';
+        ChildRoot += std::to_string(Util::getProcessId()) + '-' + Util::rng::getHexString(8) + '/';
 
         LOG_DBG("Normalizing childroot: " << ChildRoot);
         ChildRoot = Poco::Path(ChildRoot).makeDirectory().makeAbsolute().toString();
@@ -3659,7 +3663,9 @@ int COOLWSD::innerMain()
     // allocate port & hold temporarily.
     std::shared_ptr<ServerSocket> serverPort = Server->findClientPort();
 
+#if !MOBILEAPP
     TmpFontDir = ChildRoot + JailUtil::CHILDROOT_TMP_INCOMING_PATH;
+#endif
 
     // Start the internal prisoner server and spawn forkit,
     // which in turn forks first child.
@@ -3879,8 +3885,10 @@ int COOLWSD::innerMain()
 
     SigUtil::addActivity("shutting down");
 
+#if !MOBILEAPP
     // Lots of polls will stop; stop watching them first.
     SocketPoll::PollWatchdog.reset();
+#endif
 
     // Stop the listening to new connections
     // and wait until sockets close.
@@ -3904,7 +3912,6 @@ int COOLWSD::innerMain()
 
     // atexit handlers tend to free Admin before Documents
     LOG_INF("Exiting. Cleaning up lingering documents.");
-#if !MOBILEAPP
     if (remoteFontConfigThread)
     {
         LOG_DBG("Stopping remote font config thread");
@@ -3918,7 +3925,6 @@ int COOLWSD::innerMain()
                 "now.");
         SigUtil::requestShutdown();
     }
-#endif
 
     SigUtil::addActivity("wait save & close");
 
@@ -4116,10 +4122,6 @@ void COOLWSD::cleanup([[maybe_unused]] int returnValue)
 
 int COOLWSD::main(const std::vector<std::string>& /*args*/)
 {
-#if MOBILEAPP && !defined IOS
-    SigUtil::resetTerminationFlags();
-#endif
-
     int returnValue = EXIT_SOFTWARE;
 
     try {
